Add files via upload
Implement a function that adds two numbers together and returns their sum in binary. The conversion can be done before, or after the addition.

The binary number returned should be a string.

1, 1 --> "10" (1 + 1 = 2 in decimal or 10 in binary)
5, 9 --> "1110" (5 + 9 = 14 in decimal or 1110 in binary)

Binary
Fundamentals
Ad
8 Reasons Why Codewarriors Practice Coding with Codewars
Not everyone trains the same. Discover new ways to leverage Codewars in your education and career.
Solution

1

function addBinary(a,b) {

2

  return (a + b).toString(2)

3

}